review-guidelines: Update signal example
Disposing a GObject - which is what Clutter.Actor.destroy() does - will
disconnect all handlers that are connected to any of that object's
signals.

That is, the current example actually shows a case where it is perfectly
fine to *not* disconnect signals explicitly. Replace it with an example
where disconnecting the signal is strictly necessary.