extensions(overview): clarify the nature of `version`
Clarify the meaning and use of the `version` field, explicitly stating
that it is *not* actually the extensions version, but actually the
submission version.

Ideally, we should stop exposing this field at all to developers or
users, since it can only be set to the exact version that e.g.o expects
and has proven to only really cause confusion.