extensions: add advanced documentation for debugging
Instructions for debugging an extension are occasionally asked for, even
though it's usually not what a developer wants.

Add advanced documentation for running GNOME Shell through `gdb`,
documenting only the aspects specific to GJS and GNOME Shell. This
should satisfy developers who know they need to do this, while
encouraging everyone else to make the obvious decision not to.